pyproject.toml
[project]
name = "automl"
version = "0.1.0"
description = "kaggle-automl using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "arxiv>=2.1.3",
    "crewai[tools]>=0.86.0,<1.0.0",
    "huggingface-hub>=0.27.0",
    "kaggle>=1.6.17",
    "python-dotenv>=1.0.0"
]

[project.scripts]
rapidml = "automl.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
